% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_taxa_table.R
\name{convert_to_taxa_table}
\alias{convert_to_taxa_table}
\title{Create a phyloseq taxon table}
\usage{
convert_to_taxa_table(OtuTable, TAXON)
}
\arguments{
\item{OtuTable}{produced by convert_to_otu_table}

\item{TAXON}{taxonomic level (First letter must be caps). It should be the same taxonomic level
that its corresponding otu table is at.}
}
\value{
phyloseq formatted taxonomic table
}
\description{
Extracts taxonomic information from the otu_table and ncbi taxonomy file to
to create phyloseq formatted tax_table. ncbi taxonomy file is provided with the package
taxonomy.tsv.gz
}

