% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phyloseq_class.R
\name{create_phyloseq_class}
\alias{create_phyloseq_class}
\title{Creates a phyloseq class corresponding.}
\usage{
create_phyloseq_class(merge_table, meta_table = NULL, taxon = "species")
}
\arguments{
\item{merge_table}{Path to merged_table.}

\item{meta_table}{Path to metadata table with samples as rows and columns as metadata.}

\item{taxon}{A taxonomic level (phylum, class, order, family, species). Default is "species".}
}
\value{
phyloseq_class object
}
\description{
Creates a phyloseq class with taxonomy table, otu table from given merged table and metadata file.
}

