% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{empirical_power}
\alias{empirical_power}
\title{Calculation of empirical power, also known as acceptance rate,
for a list of estimated effects}
\usage{
empirical_power(results, population_effect)
}
\arguments{
\item{results}{List of estimated effects bounds in format
( lower bound, upper bound )}

\item{population_effect}{Effect of the population}
}
\value{
Empirical power, in range 0-1
}
\description{
Calculation of empirical power, also known as acceptance rate,
for a list of estimated effects
}
\examples{
rejection_rate(list(c(1,3), c(-1, 4), c(1,4)), 3.2)
}
