% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_n.R
\name{get_n_for_prespecified_error}
\alias{get_n_for_prespecified_error}
\title{Get ideal number of iterations for a Monte Carlo simulation}
\usage{
get_n_for_prespecified_error(
  required_error,
  fun0,
  fun2,
  tolerance = 0.1,
  confidence_interval = 0.95,
  initial_n = 10000
)
}
\arguments{
\item{required_error}{Required error}

\item{fun0}{Generation function}

\item{fun2}{Escalating function}

\item{tolerance}{tolerance for the error}

\item{confidence_interval}{confidence interval for the error}

\item{initial_n}{optional initial value for n}
}
\value{
Number of iterations
}
\description{
Get ideal number of iterations for a Monte Carlo simulation
}
\examples{

fun0 <- function() {
 x <- runif(metacomp::get_n_iterations())
 y <- runif(metacomp::get_n_iterations())
 ifelse(sqrt(x^2 + y^2) <= 1, 1, 0)
}

fun1 <- function(results) {
  4 * sum(results[results == 1]) / metacomp::get_n_iterations()
}

fun2 <- function(results) {
  16 * var(results)
}

x <- get_n_for_prespecified_error(0.01, fun0, fun2)

}
