% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{get_n_iterations}
\alias{get_n_iterations}
\title{Get the number of iterations the algorithm runs}
\usage{
get_n_iterations()
}
\value{
number of iterations
}
\description{
Get the number of iterations the algorithm runs
}
\examples{
get_n_iterations()
}
