% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hedges_meta}
\alias{hedges_meta}
\title{Estimated treatment effect and
lower and upper confidence interval limits
using Hedges method (Hedges & Olkin, 1985),
implementation from package Meta.}
\usage{
hedges_meta(
  sample_size,
  mean_e,
  mean_c,
  sd_e,
  n_experiments,
  distribution = "normal"
)
}
\arguments{
\item{sample_size}{Number of observations}

\item{mean_e}{Estimated mean in experimental group}

\item{mean_c}{Estimated mean in control group}

\item{sd_e}{Standard deviation}

\item{n_experiments}{Number of experiments}
}
\value{
List of effect size, lower and upper intervals
}
\description{
For one experiment or study the individual
effect is calculated, for more than one
the overall effect is calculated
}
\examples{
hedges_meta(10, 6, 5, 3, 3)
}
