% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_edge_assignments.R
\name{load_edge_assignments}
\alias{load_edge_assignments}
\title{Efficiently loads a BWA (or other EDGE-like taxonomic assignment) tables from a list
of files. Outputs a named list of assignments.}
\usage{
load_edge_assignments(filepath, type)
}
\arguments{
\item{filepath}{the path to tab delimited, two-column file whose first column is a project_id
(which will be used to name this assignment) and the second column is the assignment filename.}

\item{type}{the type of assignments to be loaded. Following types are recognized: 'bwa',
'diamond', 'gottcha', 'gottcha2', 'kraken', 'metaphlan', and 'pangia'.}
}
\value{
a list of all read assignments.
}
\description{
Efficiently loads a BWA (or other EDGE-like taxonomic assignment) tables from a list
of files. Outputs a named list of assignments.
}
