% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gottcha_assignment.R
\name{load_gottcha_assignment}
\alias{load_gottcha_assignment}
\title{Efficiently loads the GOTTCHA (or other EDGE-like taxonomic assignment) table from a file.
An assumption has been made: since EDGE tables are generated in automated fashion,
they should be properly formatted, so the code doesn't check for any inconsistencies except
for the input file existence. This implementation fully relies on the read.table function
from data.table package.}
\usage{
load_gottcha_assignment(filepath)
}
\arguments{
\item{filepath}{The EDGE-generated tab-delimeted taxonomy assignment file.}
}
\value{
the data frame corresponding to the EDGE's taxonomic assignment file.
}
\description{
Efficiently loads the GOTTCHA (or other EDGE-like taxonomic assignment) table from a file.
An assumption has been made: since EDGE tables are generated in automated fashion,
they should be properly formatted, so the code doesn't check for any inconsistencies except
for the input file existence. This implementation fully relies on the read.table function
from data.table package.
}
\examples{
DF_GOTTCHA_STRAIN <- load_assignment( "tests/test_data/248/allReads-gottcha-strDB-b.list.txt" )
}

