% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kraken_assignments.R
\name{load_kraken_assignments}
\alias{load_kraken_assignments}
\title{Efficiently loads a GOTTCHA (or other EDGE-like taxonomic assignment) tables from a list
of files. Outputs a named list of assignments.}
\usage{
load_kraken_assignments(filepath)
}
\arguments{
\item{filepath}{A path to tab delimeted, two-column file whose first column is a project_id
(which will be used to name this assignment) and the second column is the assignment filename.}
}
\value{
a list of all read assignments.
}
\description{
Efficiently loads a GOTTCHA (or other EDGE-like taxonomic assignment) tables from a list
of files. Outputs a named list of assignments.
}
