% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pangia_assignment.R
\name{load_pangia_assignment}
\alias{load_pangia_assignment}
\title{Efficiently loads a PANGIA (or other EDGE-like taxonomic assignment) table from a file.
An assumption has been made -- since PANGIA/EDGE tables are generated in an automated fashion,
they should be properly formatted -- thus the code doesn't check for any inconsistencies except
for the very file existence. This implementation fully relies on the read.table function
from data.table package gaining performance over traditional R techniques.}
\usage{
load_pangia_assignment(filepath)
}
\arguments{
\item{filepath}{A path to EDGE-generated tab-delimeted PANGIA taxonomy assignment file.}
}
\value{
a data frame containing four columns: LEVEL, TAXA, COUNT, and ABUNDANCE,
        representing taxonomically anchored sequences from the sample.
}
\description{
Efficiently loads a PANGIA (or other EDGE-like taxonomic assignment) table from a file.
An assumption has been made -- since PANGIA/EDGE tables are generated in an automated fashion,
they should be properly formatted -- thus the code doesn't check for any inconsistencies except
for the very file existence. This implementation fully relies on the read.table function
from data.table package gaining performance over traditional R techniques.
}
