% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gottcha_columns.R
\name{merge_gottcha_columns}
\alias{merge_gottcha_columns}
\title{Merges two EDGE's taxonomical assignments. The input data frames are assumed to have the
LEVEL column first, TAXA column second, following by a single column of abundance values.
Outputs a table whose first two columns, LEVEL and TAXA, contain a union of these
from both tables and whose second and third column contain the given abundances.}
\usage{
merge_gottcha_columns(assignment1, assignment2)
}
\arguments{
\item{assignment1}{The gottcha-like, three-column assignment table.}

\item{assignment2}{The gottcha-like, three-column assignment table.}
}
\value{
a merged table.
}
\description{
Merges two EDGE's taxonomical assignments. The input data frames are assumed to have the
LEVEL column first, TAXA column second, following by a single column of abundance values.
Outputs a table whose first two columns, LEVEL and TAXA, contain a union of these
from both tables and whose second and third column contain the given abundances.
}

