% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metaphlan_assignments.R
\name{merge_metaphlan_assignments}
\alias{merge_metaphlan_assignments}
\title{Merges two or more Metaphlan-like taxonomical assignments. The input data frames are assumed to
have the following columns: LEVEL, TAXA, and ROLLUP -- these will be used in the merge
procedure, all other columns will be ignored.}
\usage{
merge_metaphlan_assignments(assignments)
}
\arguments{
\item{assignments}{A named list of assignments (the list element's name will be used as a
resulting data frame column name).}
}
\value{
A merged table, which is a data frame whose rows are taxonomical ids and columns are
         the input assignments ids.
}
\description{
Merges two or more Metaphlan-like taxonomical assignments. The input data frames are assumed to
have the following columns: LEVEL, TAXA, and ROLLUP -- these will be used in the merge
procedure, all other columns will be ignored.
}
