% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{montecarlo_meta}
\alias{montecarlo_meta}
\title{Title}
\usage{
montecarlo_meta(
  matrizComb,
  random_generation_function,
  metric_calculation_function
)
}
\arguments{
\item{fun1}{}
}
\description{
Title
}
