% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot}
\alias{plot}
\title{Plot a simple simulation with a single parameter}
\usage{
plot(matrizComb, method, metric, parameter, ...)
}
\arguments{
\item{parameter}{}
}
\value{
a plot ?
}
\description{
Plot a simple simulation with a single parameter
}
