% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bwa_assignment.R
\name{plot_bwa_assignment}
\alias{plot_bwa_assignment}
\title{Generates a single column ggplot for a taxonomic assignment table and also outputs a PDF.}
\usage{
plot_bwa_assignment(assignment, level, plot_title, column_title, filename)
}
\arguments{
\item{assignment}{The BWA-like assignment table.}

\item{level}{The taxonomic level to plot (i.e., family, strain, etc...).}

\item{plot_title}{The plot title, e.g., "Project XX, Run YY".}

\item{column_title}{The column title.}

\item{filename}{The PDF file name mask.}
}
\value{
the ggplot2 plot.
}
\description{
This implementation is built upon ggplot geom_tile.
}
