% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_choose_method}
\alias{plot_choose_method}
\title{Plot metric equivalences between two methods}
\usage{
plot_choose_method(equivalences, ...)
}
\arguments{
\item{equivalences}{metric equivalences between two methods}
}
\value{
a plot
}
\description{
Plot metric equivalences between two methods
}
