% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_compare_methods}
\alias{plot_compare_methods}
\title{Plot a comparison of methods}
\usage{
plot_compare_methods(matrix_comb, metric, parameter, ...)
}
\arguments{
\item{matrix_comb}{methods and comparison information}

\item{metric}{metric used in comparison to plot in vertical axis}

\item{parameter}{parameter of interest to plot in horizontal axis}
}
\value{
Graphical plot
}
\description{
Plot a comparison of methods
}
