% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gottcha_as_column.R
\name{plot_gottcha_as_column}
\alias{plot_gottcha_as_column}
\title{Generates a single column ggplot for a taxonomic assignment table.}
\usage{
plot_gottcha_as_column(assignment_df, taxonomy_level, db_desc)
}
\arguments{
\item{assignment_df}{The gottcha-like assignment table.}

\item{taxonomy_level}{The level which need to be plotted.}

\item{db_desc}{The column title, something like "ottcha-strDB-v", etc.}
}
\value{
the ggplot2 plot.
}
\description{
This implementation is built upon ggplot tile.
}

