% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gottcha_columns.R
\name{plot_gottcha_columns}
\alias{plot_gottcha_columns}
\title{Generates a single column ggplot for a taxonomic assignment table.}
\usage{
plot_gottcha_columns(assignment, taxonomy_level)
}
\arguments{
\item{assignment}{The gottcha-like merged assignment table.}

\item{taxonomy_level}{The level which need to be plotted.}
}
\value{
the ggplot2 plot.
}
\description{
This implementation...
}

