% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmds_analysis.R
\name{plot_nmds_analysis}
\alias{plot_nmds_analysis}
\title{Nonmetric Multidimensional Scaling}
\usage{
plot_nmds_analysis(filepath, LEVEL = NULL, metadata, shape_grp, col_grp)
}
\arguments{
\item{filepath}{path to merged_table (a tsv file) containing samples as column}

\item{LEVEL}{A taxonomic level (phylum, class, order, family, species) to create OTU table.}

\item{metadata}{A tab delimited metadata table with samples as rows and variables/properties as columns, no column header for sample.}

\item{shape_grp}{header of the metadata to refer for shape type.}

\item{col_grp}{header of the metadata to refer for color type.}
}
\value{
A cowplot object
}
\description{
Creates a cowplot object with annotated nmds plot
}

