% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sample_generator}
\alias{sample_generator}
\title{lower and upper confidence interval limits}
\usage{
sample_generator(
  n_experiments,
  sample_size,
  mean_e,
  sd_e,
  mean_c,
  sd_c,
  distribution
)
}
\arguments{
\item{n_experiments}{Number of experiments}

\item{sample_size}{Number of observations}

\item{mean_e}{Estimated mean in experimental group}

\item{sd_e}{Standard deviation in experimental group}

\item{mean_c}{Estimated mean in control group}

\item{sd_c}{Standard deviation in control group}

\item{distribution}{Distibution model}
}
\value{
List of generated samples for
experimental and control group
}
\description{
lower and upper confidence interval limits
}
\examples{
sample_generator(12, 50, 6, 3, 5, 3, "normal")
}
