% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{set_n_iterations}
\alias{set_n_iterations}
\title{Set the number of iterations the algorithm runs}
\usage{
set_n_iterations(val)
}
\arguments{
\item{val}{Number of iterations}
}
\description{
Set the number of iterations the algorithm runs
}
\examples{
set_n_iterations(200)
}
