% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_alpha_div.R
\name{summary_alpha_div}
\alias{summary_alpha_div}
\title{Alpha diversity}
\usage{
summary_alpha_div(pseq_class)
}
\arguments{
\item{pseq_class}{phyloseq class}
}
\value{
A cowplot object
}
\description{
A plot object with alpha diversity plots
}

