% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_heatmaps.R
\name{summary_heatmaps}
\alias{summary_heatmaps}
\title{Summarize taxonomic data on different taxonomic levels.
Creates a pdf file and outputs a cowplot object with summary heatmaps (derived from phyloseq) of specific taxonomic order (when given) or all possible taxonomic order}
\usage{
summary_heatmaps(filepath, LEVEL = NULL, out_pdf = "taxa_heatmap.pdf")
}
\arguments{
\item{filepath}{path to merged_table containing samples as column}

\item{LEVEL}{A taxonomic level (phylum, class, order, family, species) to create OTU table.}

\item{out_pdf}{default = taxa_heatmap.pdf The PDF file name mask.}
}
\value{
A cowplot object
}
\description{
Summarize taxonomic data on different taxonomic levels.
Creates a pdf file and outputs a cowplot object with summary heatmaps (derived from phyloseq) of specific taxonomic order (when given) or all possible taxonomic order
}

