% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pheatmap.R
\name{summary_pheatmap}
\alias{summary_pheatmap}
\title{Summarize taxonomic assignments into a pheatmap.}
\usage{
summary_pheatmap(filepath, LEVEL = NULL, ...)
}
\arguments{
\item{filepath}{path to merged_table (a tsv file) containing samples as column}

\item{LEVEL}{A taxonomic level (phylum, class, order, family, species) to create OTU table}

\item{...}{origingal parameters from pheatmap}
}
\value{
A cowplot object derived from pheatmap gtable
}
\description{
Returns pheatmap(s) for specific taxonomic order (when given) or all possible taxonomic order when LEVEL is not specified.
}

