% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_phylo_heatmap.R
\name{summary_phylo_heatmap}
\alias{summary_phylo_heatmap}
\title{Summarize taxonomic assignments.}
\usage{
summary_phylo_heatmap(filepath, LEVEL = NULL)
}
\arguments{
\item{filepath}{path to merged_table (a tsv file) containing samples as column}

\item{LEVEL}{A taxonomic level (phylum, class, order, family, species) to create OTU table.}
}
\value{
A cowplot object
}
\description{
Creates a cowplot object with summary heatmaps (derived from phyloseq) of
specific taxonomic order (when given) or all possible taxonomic order
}

