% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy}
\alias{accuracy}
\title{Calculation of accuracy, as in the empirical probability of the
meta-analysis method confidence interval (1 − α) containing the
population value δ}
\usage{
accuracy(results, d)
}
\arguments{
\item{results}{List of estimated effects in format
( effect, lower bound, upper bound )}

\item{d}{Effect of the population}
}
\value{
Accuracy, in range 0-1
}
\description{
Calculation of accuracy, as in the empirical probability of the
meta-analysis method confidence interval (1 − α) containing the
population value δ
}
\examples{
accuracy(list(c(2,1,3), c(3,-1, 4), c(2, 1,4)), 3.2)
accuracy(estimated_effects)
}
