% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_otu_table.R
\name{convert_to_otu_table}
\alias{convert_to_otu_table}
\title{Create Phyloseq format otu table}
\usage{
convert_to_otu_table(filepath, TAXON)
}
\arguments{
\item{filepath}{to merge table containig samples as columns and taxa as rows.}

\item{TAXON}{A taxonomic level (phylum, class, order, family, species) to create OTU table.
It has to be all small letters.}
}
\value{
phyloseq otu table
}
\description{
Synthesize phyloseq format otu table from merged table for given taxonomic level.
OTU here corresponds to the taxonomic level. This implementation is built upon phyloseq. We recommend not to use this for strain
level classification as creation of taxa_table doesn't work with strain level or
plasmid sequences.
}

