% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide.R
\name{decide}
\alias{decide}
\title{Title}
\usage{
decide(
  methods,
  metric,
  min_n_experiments,
  max_n_experiments,
  min_sample_size,
  max_sample_size,
  mean_e,
  mean_c,
  sd_e,
  sd_c,
  distribution = "normal",
  parameter_of_interest = ""
)
}
\description{
Title
}
