% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{empirical_power}
\alias{empirical_power}
\title{Calculation of empirical power, also known as acceptance rate,
for a list of estimated effects}
\usage{
empirical_power(results)
}
\arguments{
\item{results}{List of estimated effects in format
( effect, lower bound, upper bound )}
}
\value{
Empirical power, in range 0-1
}
\description{
Calculation of empirical power, also known as acceptance rate,
for a list of estimated effects
}
\examples{
rejection_rate(list(c(2,1,3), c(3,-1, 4), c(2, 1,4)))
rejection_rate(estimated_effects)
}
