% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hedges_meta}
\alias{hedges_meta}
\title{Estimated treatment effect and
lower and upper confidence interval limits
using Hedges method (Hedges & Olkin, 1985):}
\usage{
hedges_meta(sample_size, mean_e, mean_c, sd_e, n_experiments)
}
\arguments{
\item{sample_size}{Number of observations}

\item{mean_e}{Estimated mean in experimental group}

\item{mean_c}{Estimated mean in control group}

\item{sd_e}{Standard deviation}

\item{n_experiments}{Number of experiments}
}
\value{
List of effect size, lower and upper intervals
}
\description{
d = J(N-2)(Y_e - Y_c)/S_p
}
\details{
where
d is the effect size
J (N – 2) = is the correction factor
Y is the mean of experimental (E) and control (C)
groups
Sp is the pooled standard deviation of both groups
N is the number of experimental subjects of both groups (nE + nC)

For one experiment or study the individual
effect is calculated, for more than one
the overall effect is calculated
}
\examples{
hedges_meta(10, 6, 5, 3, 3)
}
