% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gottcha_assignments.R
\name{merge_gottcha_assignments}
\alias{merge_gottcha_assignments}
\title{Merges two GOTTCHA-like taxonomical assignments. The input data frames are assumed to have the
following columns: LEVEL, TAXA, and NORM_COV -- these will be used in the merge procedure.}
\usage{
merge_gottcha_assignments(assignment1, assignment2)
}
\arguments{
\item{assignment1}{The gottcha-like assignment table.}

\item{assignment2}{The gottcha-like assignment table.}
}
\value{
a merged table.
}
\description{
Merges two GOTTCHA-like taxonomical assignments. The input data frames are assumed to have the
following columns: LEVEL, TAXA, and NORM_COV -- these will be used in the merge procedure.
}

