% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{parametric_rr_meta}
\alias{parametric_rr_meta}
\title{Estimated treatment effect and
lower and upper confidence interval limits
using the Parametric Response Ratio method
(Hedges et al, 1999), also known as
Ratio of Means,
implementation from package Meta.}
\usage{
parametric_rr_meta(
  sample_size,
  mean_e,
  sd_e,
  mean_c,
  sd_c,
  n_experiments,
  distribution = "normal",
  return = NULL
)
}
\arguments{
\item{sample_size}{Number of observations}

\item{mean_e}{Estimated mean in experimental group}

\item{sd_e}{Standard deviation in experimental group}

\item{mean_c}{Estimated mean in control group}

\item{sd_c}{Standard deviation in control group}

\item{n_experiments}{Number of experiments}

\item{distribution}{Distibution model}

\item{return}{Return mode}
}
\value{
Bounds, population effect or sample size based
on parameter return
}
\description{
For one experiment or study the individual
effect is calculated, for more than one
the overall effect is calculated
}
\examples{
hedges_meta(10, 10, 3, 8, 3, 12, "population effect")
}
