% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{parametric_rr_meta}
\alias{parametric_rr_meta}
\title{Estimated treatment effect and
lower and upper confidence interval limits
using the Parametric Response Ratio method
\link{Hedges, L.; et al; 1999}, also known as
Ratio of Means:}
\usage{
parametric_rr_meta(sample_size, mean_e, mean_c, sd_e, n_experiments)
}
\arguments{
\item{sample_size}{Number of observations}

\item{mean_e}{Estimated mean in experimental group}

\item{mean_c}{Estimated mean in control group}

\item{sd_e}{Standard deviation}

\item{n_experiments}{Number of experiments}
}
\value{
List of effect size, lower and upper intervals
}
\description{
RR = Y_e / Y_c
}
\details{
where
RR is the response ratio
Y is the mean of experimental (E) and control (C)
groups

which is normalized as

L= Ln(RR)

where L is the natural logarithm of the response
ratio

For one experiment or study the individual
effect is calculated, for more than one
the overall effect is calculated
}
\examples{
parametric_rr_meta(10, 6, 5, 3, 3)
}
