% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_assignment.R
\name{plot_merged_assignment}
\alias{plot_merged_assignment}
\title{Generates a single column ggplot for a taxonomic assignment table.}
\usage{
plot_merged_assignment(assignment, taxonomy_level, title, filename)
}
\arguments{
\item{assignment}{The gottcha-like merged assignment table.}

\item{taxonomy_level}{The level which need to be plotted.}

\item{title}{The plot title.}

\item{filename}{The PNG file name.}
}
\value{
the ggplot2 plot.
}
\description{
This implementation...
}

