% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rejection_rate}
\alias{rejection_rate}
\title{Calculation of rejection rate for a list of estimated effects}
\usage{
rejection_rate(results)
}
\arguments{
\item{results}{List of estimated effects in format
\link{ effect, lower bound, upper bound }}
}
\value{
Rate of rejection, in range 0-1
}
\description{
Calculation of rejection rate for a list of estimated effects
}
\examples{
rejection_rate(list(c(2,1,3), c(3,-1, 4), c(2, 1,4)))
rejection_rate(estimated_effects)
}
